Create Dom module
This module is useful for indicating how far over an element a mouse has
passed. This was just enough of sore thumb in main that it validates
extraction, but I also want to use it in another project for
a thumbnail-style view, and so here we are.

I'm not super jazzed about the names. Something to represent a value
between 0 and 1 is the hard part, without resorting to something like
"percentage".