Isolate System as a module
There's a lot about System that's self-contained: the SVG calls, the
path, expansion of the grammar, the type itself. It's just not easily
to performantly draw that thing. With this in place, hopefully I can
start chipping away more concretely at the path generation stuff.

This also extracts Colorscheme into a module.