Report Pointer position
"Dom" was too vague for what that module was trying to do. Plus, what we
really want to do is keep track of mouse or finger position over
a surface. It makes less sense to wrap up the `Html.Events` and `Svg.Events`
packages than it does to provide a decoder that can grab mouse position.

In renaming those mouse functions under a `Pointer` module, I'd love to
adapt this to something that could also get touch event positions.
That's a little harder to do, because we need to batch potentially many
offsets into a list. It would be nice to take care of event listeners as
well. So rather than doing:

- Html.Events.on "mousemove" |> Pointer.progress msg
- Html.Events.on "touchmove" |> Pointer.progress msg

There could be something wrapping the pointer events api:

- Html.Events.on "pointermove" |> Pointer.progress msg

Where the polyfill for the event happens by including javascript on the
page. Another option might be an Elm-level pointer concept:

- Pointer.onOver msg
- Pointer.onEnter msg
- Pointer.onMove msg
- Pointer.onUp msg
- Pointer.onOut msg

Whatever!