Allow for radio items to be checked
Previously there wasn't a "name" attribute on each radio input, which
allowed multiple radio items to be "selected" at a time. There also
wasn't a "checked" attribute, which is how we can programatically
control which value is selected.

This change introduces those two properties, though for now, they are
hard-coded to the "triangles" preset.