Move logic about selected item into union control
Makes sense to keep track of selected item state separate from the list
of options. Though this is turning into a lot of stuff to pass to the
`union` control!

To follow the elm-sortable-table approach, a signature like this might
make sense:

```
union : UnionOptions -> UnionState -> Choices
```

Where `UnionOptions` holds onto the Msg and the Label, state is
the selected type, and choices is an alias for `List ( String, choice
)`. More experimentation is necessary!