Un-nest top-level model
I had some distinction between a loading page and a loaded page, where
the model was pending, or was created with a random colorscheme. That
random call is what prevents loading of the app initially.

This change makes the model available from the first render, but
replaces a black-and-white colorscheme with the randomly loaded one.

This was also an opportunity to join the `Preset` and `Image` records
together under the `Model` record.

The end result is that the update function is very compact now. Less
need to reach into deeply nested structures, and build up an
intermediate result.